// src/functional-modules/evm/interfaces/rs-kies/block-data-node-rs.interface.ts
export interface IBlockDataFromEvm {
    baseFeePerGas: string;
    difficulty: string;
    extraData: string;
    gasLimit: string;
    gasUsed: string;
    hash: string;
    logsBloom: string;
    miner: string;
    mixHash: string;
    nonce: string;
    number: string;
    parentHash: string;
    receiptsRoot: string;
    sha3Uncles: string;
    size: string;
    stateRoot: string;
    timestamp: string;
    transactions: string[];
    transactionsRoot: string[];
    uncles: string[];
}